requirements.txt
boto3==1.4.4
botocore==1.5.46
docutils==0.13.1
ExifRead==2.1.2
jmespath==0.9.2
olefile==0.44
piexif==1.0.12
Pillow==4.1.1
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
